+++ R/argument_validation.R
@@ -26,7 +26,7 @@
 #'
 #' @export
 
-match_arg <- function (arg, choices, several.ok = FALSE, none.ok = FALSE) {
+match_arg <- function(arg, choices, several.ok = FALSE, none.ok = FALSE) {
   checkmate::assert_character(arg)
   checkmate::assert_character(choices)
   checkmate::assert_flag(several.ok)
@@ -81,18 +81,21 @@ match_arg <- function (arg, choices, several.ok = FALSE, none.ok = FALSE) {
 #' @export
 
 check_covariance_matrix <- function(
-    x, dim = NULL, tolerance = sqrt(.Machine$double.eps)
-  ) {
+    x, dim = NULL, tolerance = sqrt(.Machine$double.eps)) {
   checkmate::assert_number(tolerance, lower = 0)
   res <- checkmate::check_matrix(x, mode = "numeric")
-  if (!isTRUE(res))
+  if (!isTRUE(res)) {
     return(res)
-  if (nrow(x) != ncol(x))
+  }
+  if (nrow(x) != ncol(x)) {
     return("Must be square")
-  if (any(abs(x - t(x)) > tolerance))
+  }
+  if (any(abs(x - t(x)) > tolerance)) {
     return("Must be symmetric")
-  if (any(eigen(x)$value < -tolerance))
+  }
+  if (any(eigen(x)$value < -tolerance)) {
     return("Must have positive eigenvalues only")
+  }
   if (!is.null(dim)) {
     checkmate::assert_count(dim, positive = TRUE)
     if (nrow(x) != dim) {
@@ -138,20 +141,24 @@ test_covariance_matrix <- checkmate::makeTestFunction(
 #' @export
 
 check_correlation_matrix <- function(
-    x, dim = NULL, tolerance = sqrt(.Machine$double.eps)
-  ) {
+    x, dim = NULL, tolerance = sqrt(.Machine$double.eps)) {
   checkmate::assert_number(tolerance, lower = 0)
   res <- checkmate::check_matrix(x, mode = "numeric")
-  if (!isTRUE(res))
+  if (!isTRUE(res)) {
     return(res)
-  if (nrow(x) != ncol(x))
+  }
+  if (nrow(x) != ncol(x)) {
     return("Must be square")
-  if (any(abs(x - t(x)) > tolerance))
+  }
+  if (any(abs(x - t(x)) > tolerance)) {
     return("Must be symmetric")
-  if (any(abs(diag(x) - 1) > tolerance))
+  }
+  if (any(abs(diag(x) - 1) > tolerance)) {
     return("Must have ones on the diagonal")
-  if (any(x < -1 | x > 1))
+  }
+  if (any(x < -1 | x > 1)) {
     return("Must have values between -1 and 1")
+  }
   if (!is.null(dim)) {
     checkmate::assert_count(dim, positive = TRUE)
     if (nrow(x) != dim) {
@@ -197,18 +204,21 @@ test_correlation_matrix <- checkmate::makeTestFunction(
 #' @export
 
 check_transition_probability_matrix <- function(
-    x, dim = NULL, tolerance = sqrt(.Machine$double.eps)
-  ) {
+    x, dim = NULL, tolerance = sqrt(.Machine$double.eps)) {
   checkmate::assert_number(tolerance, lower = 0)
   res <- checkmate::check_matrix(x, mode = "numeric")
-  if (!isTRUE(res))
+  if (!isTRUE(res)) {
     return(res)
-  if (nrow(x) != ncol(x))
+  }
+  if (nrow(x) != ncol(x)) {
     return("Must be square")
-  if (any(x < 0 | x > 1))
+  }
+  if (any(x < 0 | x > 1)) {
     return("Must have values between 0 and 1")
-  if (any(abs(rowSums(x) - 1) > tolerance))
+  }
+  if (any(abs(rowSums(x) - 1) > tolerance)) {
     return("Must have row sums equal to 1")
+  }
   if (!is.null(dim)) {
     checkmate::assert_count(dim, positive = TRUE)
     if (nrow(x) != dim) {
@@ -253,17 +263,19 @@ test_transition_probability_matrix <- checkmate::makeTestFunction(
 #' @export
 
 check_probability_vector <- function(
-    x, len = NULL, tolerance = sqrt(.Machine$double.eps)
-  ) {
+    x, len = NULL, tolerance = sqrt(.Machine$double.eps)) {
   checkmate::assert_number(tolerance, lower = 0)
   res1 <- checkmate::check_atomic_vector(x, any.missing = FALSE, len = len)
-  if (!isTRUE(res1))
+  if (!isTRUE(res1)) {
     return(res1)
+  }
   res2 <- checkmate::check_numeric(x, lower = 0, upper = 1)
-  if (!isTRUE(res2))
+  if (!isTRUE(res2)) {
     return(res2)
-  if (abs(sum(x) - 1) > tolerance)
+  }
+  if (abs(sum(x) - 1) > tolerance) {
     return("Must add up to 1")
+  }
   return(TRUE)
 }