+++ R/package_helpers.R
@@ -12,11 +12,12 @@
 #' Either \code{TRUE} to have curly brackets around the package name (default)
 #' or \code{FALSE} if not.
 #'
+#' @export
+#'
 #' @return
 #' No return value, but it runs \code{\link[usethis]{use_logo}} in the end.
 
 basic_package_sticker <- function(package_name, brackets = TRUE) {
-
   ### input checks
   checkmate::assert_string(package_name)
   checkmate::assert_flag(brackets)
@@ -44,7 +45,8 @@ basic_package_sticker <- function(package_name, brackets = TRUE) {
   sticker_file <- hexSticker::sticker(
 
     ### image
-    subplot = ggplot2::ggplot() + ggplot2::theme_void(),
+    subplot = ggplot2::ggplot() +
+      ggplot2::theme_void(),
     s_x = 1,
     s_y = 1,
     s_width = 2,
@@ -84,11 +86,109 @@ basic_package_sticker <- function(package_name, brackets = TRUE) {
     filename = filename,
     asp = 1,
     dpi = 300
-  ); sticker_file
+  )
+  sticker_file
 
   ### message
   message("path:", filename)
 
   ### set sticker
   usethis::use_logo(img = filename)
+}
+
+#' Handling of an unexpected error
+#'
+#' @description
+#' This function reacts to an unexpected error by throwing an error and linking
+#' to a GitHub issues site with the request to submit an issue.
+#'
+#' @param msg
+#' A \code{character}, an error message.
+#' @param issue_link
+#' A \code{character}, the URL to a GitHub issues site.
+#'
+#' @export
+#'
+#' @return
+#' No return value, but it throws an error.
+
+unexpected_error <- function(
+    msg = "We are sorry, an unexpected error occured.",
+    issue_link = "https://github.com/loelschlaeger/oeli/issues") {
+  checkmate::assert_string(msg, min.chars = 1)
+  checkmate::assert_string(
+    issue_link,
+    pattern = "^https://github.com/[[:alpha:]]*/[[:alpha:]]*/issues$"
+  )
+  cli::cli_abort(
+    c(
+      msg,
+      "i" = paste("Please submit an issue here:", cli::style_hyperlink(issue_link, issue_link))
+    ),
+    call = NULL
+  )
+}
+
+#' Using development packages when working with \code{{renv}}
+#'
+#' @description
+#' This function creates a file that loads development packages so that
+#' \code{{renv}} can detect and write them to the lockfile.
+#'
+#' @param packages
+#' A \code{character} \code{vector} of package names.
+#' @param file_name
+#' A single \code{character}, the name for the \code{.R} file.
+#'
+#' @export
+#'
+#' @return
+#' No return value, but it writes a file to the project root and adds an
+#' entry to the \code{.Rbuildignore} file.
+
+renv_development_packages <- function(
+    packages = c("covr", "devtools", "DT", "markdown", "R.utils", "yaml"),
+    file_name = "development_packages") {
+  is_package <- tryCatch(
+    rprojroot::find_package_root_file(),
+    error = function(e) FALSE
+  )
+  if (isFALSE(is_package)) {
+    cli::cli_abort(
+      c("Please check that you are inside a package directory."),
+      call = NULL
+    )
+  } else {
+    checkmate::assert_names(packages)
+    checkmate::assert_character(file_name)
+    file_path <- rprojroot::find_package_root_file(paste0(file_name, ".R"))
+    if (file.exists(file_path)) {
+      overwrite <- user_confirm(
+        question = paste(
+          "The file", file_path, "already exists, may I overwrite?"
+        ),
+        default = FALSE
+      )
+      if (!overwrite) {
+        cli_inform(
+          "Okay, nothing is changed."
+        )
+        return(invisible(NULL))
+      }
+    }
+    file_connection <- file(file_path, open = "wt")
+    for (package in packages) {
+      writeLines(
+        c(
+          paste0("if (!require(\"", package, "\", quietly = TRUE)) {"),
+          paste0("  renv::install(\"", package, "\", prompt = FALSE)"),
+          "}"
+        ),
+        con = file_connection
+      )
+    }
+  }
+  close(file_connection)
+  usethis::use_build_ignore(paste0(file_name, ".R"))
+  return(invisible(NULL))
 }