+++ R/storage_helpers.R
@@ -1,4 +1,4 @@
-#' Index R6 Object
+#' Storage R6 Object
 #'
 #' @description
 #' Provides a simple indexing interface for list elements based on R6.
@@ -40,9 +40,9 @@
 #'
 #' @return
 #' The output depends on the method:
-#' - \code{$new()} returns an \code{Index} object.
+#' - \code{$new()} returns a \code{Storage} object.
 #' - \code{$add()}, \code{$remove()}, and \code{$print()} invisibly return the
-#'   \code{Index} object (to allow for method chaining)
+#'   \code{Storage} object (to allow for method chaining)
 #' - \code{$get()} returns the requested element(s)
 #' - \code{$number()} returns an \code{integer}
 #' - \code{$indices()} return an \code{integer} \code{vector}
@@ -50,11 +50,11 @@
 #' @export
 #'
 #' @examples
-#' ### 1. Create an `Index` object:
-#' my_index <- Index$new()
+#' ### 1. Create a `Storage` object:
+#' my_storage <- Storage$new()
 #'
 #' # 2. Add elements along with identifiers:
-#' my_index$
+#' my_storage$
 #'   add(42, c("number", "rational"))$
 #'   add(pi, c("number", "!rational"))$
 #'   add("fear of black cats", c("text", "!rational"))$
@@ -62,41 +62,35 @@
 #'   add(mean, "function")
 #'
 #' # 3. What elements are stored?
-#' print(my_index)
+#' print(my_storage)
 #'
 #' # 4. Extract elements based on identifiers:
-#' my_index$get("rational")
-#' my_index$get("!rational")
-#' my_index$get(c("text", "!rational"))
-#' my_index$get("all")                        # get all elements
-#' my_index$get(c("text", "!text"))
-#' my_index$get(c("text", "!text"), logical = "or")
+#' my_storage$get("rational")
+#' my_storage$get("!rational")
+#' my_storage$get(c("text", "!rational"))
+#' my_storage$get("all") # get all elements
+#' my_storage$get(c("text", "!text"))
+#' my_storage$get(c("text", "!text"), logical = "or")
 #'
 #' # 5. Extract elements based on ids:
-#' my_index$get(ids = 4:5)
-#' my_index$get(ids = 4:5, id_names = TRUE)   # add the ids as names
-
-Index <- R6::R6Class(
-
-  classname = "Index",
+#' my_storage$get(ids = 4:5)
+#' my_storage$get(ids = 4:5, id_names = TRUE) # add the ids as names
+Storage <- R6::R6Class(
+  classname = "Storage",
   lock_class = TRUE,
   cloneable = FALSE,
-
   public = list(
 
     #' @description
-    #' initializing an \code{Index} object
+    #' initializing a \code{Storage} object
     #' @return
-    #' a new \code{Index} object
-
+    #' a new \code{Storage} object
     initialize = function() {
-
       ### elements are saved in a list
       private$elements <- list()
 
       ### ids are saved in a data.frame with identifiers in columns
       private$ids <- data.frame()
-
     },
 
     #' @description
@@ -104,19 +98,16 @@ Index <- R6::R6Class(
     #' @param x
     #' any object to be saved
     #' @return
-    #' invisibly the \code{Index} object
+    #' invisibly the \code{Storage} object
 
-    add = function(
-      x, identifier, confirm = interactive() & self$confirm,
-      missing_identifier = self$missing_identifier
-    ) {
-
+    add = function(x, identifier, confirm = interactive() & self$confirm,
+                   missing_identifier = self$missing_identifier) {
       ### input checks
       if (missing(x)) {
-        stop("please specify the object 'x' to be added", call = FALSE)
+        stop("please specify the object 'x' to be added", call. = FALSE)
       }
       if (missing(identifier)) {
-        stop("please specify at least one entry for 'identifier'", call = FALSE)
+        stop("please specify at least one entry for 'identifier'", call. = FALSE)
       }
       private$check_input(
         identifier = identifier, confirm = confirm,
@@ -138,7 +129,6 @@ Index <- R6::R6Class(
       ### add element
       private$add_element(x, identifier, missing_identifier)
       invisible(self)
-
     },
 
     #' @description
@@ -149,27 +139,20 @@ Index <- R6::R6Class(
     #' @return
     #' the selected object(s)
 
-    get = function(
-      identifier = character(), ids = integer(), logical = "and",
-      confirm = interactive() & self$confirm,
-      missing_identifier = self$missing_identifier,
-      id_names = FALSE
-    ) {
-
+    get = function(identifier = character(), ids = integer(), logical = "and",
+                   confirm = interactive() & self$confirm,
+                   missing_identifier = self$missing_identifier,
+                   id_names = FALSE) {
       ### input checks
       private$check_input(
         identifier = identifier, confirm = confirm, ids = ids,
         missing_identifier = missing_identifier, logical = logical
       )
-
-      ### check for unknown identifier
-      if (length(identifier) > 0) {
-        identifier <- private$check_identifier_known(identifier)
-      } else if (length(ids) == 0) {
+      if (length(identifier) == 0 && length(ids) == 0) {
         if (!self$hide_warnings) {
           warning(
             "please specify either 'identifier' or 'ids'",
-            call. = FALSE, immediate. = TRUE
+            call. = FALSE, immediate. = FALSE
           )
         }
         return(list())
@@ -189,7 +172,6 @@ Index <- R6::R6Class(
         identifier = identifier, ids = ids, id_names = id_names,
         logical = logical
       )
-
     },
 
     #' @description
@@ -198,26 +180,25 @@ Index <- R6::R6Class(
     #' either \code{TRUE} to shift ids when in-between elements are removed,
     #' or \code{TRUE} to keep the ids
     #' @return
-    #' invisibly the \code{Index} object
+    #' invisibly the \code{Storage} object
 
-    remove = function(
-      identifier = character(), ids = integer(), logical = "and",
-      confirm = interactive() & self$confirm,
-      missing_identifier = self$missing_identifier, shift_ids = TRUE
-    ) {
-
+    remove = function(identifier = character(), ids = integer(), logical = "and",
+                      confirm = interactive() & self$confirm,
+                      missing_identifier = self$missing_identifier, shift_ids = TRUE) {
       ### input checks
       private$check_input(
         identifier = identifier, confirm = confirm, ids = ids,
         missing_identifier = missing_identifier, logical = logical
       )
       checkmate::assert_flag(shift_ids)
-
-      ### check for unknown identifier
-      if (length(identifier) > 0) {
-        identifier <- private$check_identifier_known(identifier)
-      } else if (length(ids) == 0) {
-        stop("please specify either 'identifier' or 'ids'")
+      if (length(identifier) == 0 && length(ids) == 0) {
+        if (!self$hide_warnings) {
+          warning(
+            "please specify either 'identifier' or 'ids'",
+            call. = FALSE, immediate. = FALSE
+          )
+        }
+        return(list())
       }
 
       ### inform user about action and request confirmation
@@ -247,7 +228,6 @@ Index <- R6::R6Class(
         private$ids[ids, ] <- NA
       }
       invisible(self)
-
     },
 
     #' @description
@@ -255,21 +235,13 @@ Index <- R6::R6Class(
     #' @return
     #' an \code{integer}
 
-    number = function(
-      identifier = "all", missing_identifier = self$missing_identifier,
-      logical = "and", confirm = FALSE
-    ) {
-
+    number = function(identifier = "all", missing_identifier = self$missing_identifier,
+                      logical = "and", confirm = FALSE) {
       ### input checks
       private$check_input(
         identifier = identifier, missing_identifier = missing_identifier,
         confirm = confirm, logical = logical
       )
-
-      ### check for unknown identifier
-      if (length(identifier) > 0) {
-        identifier <- private$check_identifier_known(identifier)
-      }
 
       ### inform user about action and request confirmation
       if (confirm) {
@@ -282,7 +254,6 @@ Index <- R6::R6Class(
 
       ### perform action
       length(self$indices(identifier, confirm = confirm, logical = logical))
-
     },
 
     #' @description
@@ -290,19 +261,12 @@ Index <- R6::R6Class(
     #' @return
     #' an \code{integer} \code{vector}
 
-    indices = function(
-      identifier = "all", logical = "and",
-      confirm = interactive() & self$confirm
-    ) {
-
+    indices = function(identifier = "all", logical = "and",
+                       confirm = interactive() & self$confirm) {
+      ### input checks
       private$check_input(
         identifier = identifier, logical = logical, confirm = confirm
       )
-
-      ### check for unknown identifier
-      if (length(identifier) > 0) {
-        identifier <- private$check_identifier_known(identifier)
-      }
 
       ### inform user about action and request confirmation
       if (confirm) {
@@ -315,7 +279,6 @@ Index <- R6::R6Class(
 
       ### perform action
       private$get_ids(identifier = identifier, logical = logical)
-
     },
 
     #' @description
@@ -323,7 +286,7 @@ Index <- R6::R6Class(
     #' @param ...
     #' currently not used
     #' @return
-    #' invisibly the \code{Index} object
+    #' invisibly the \code{Storage} object
 
     print = function(...) {
       nelements <- length(private$elements)
@@ -335,13 +298,10 @@ Index <- R6::R6Class(
       }
       invisible(self)
     }
-
   ),
-
   active = list(
 
     #' @field identifier a \code{character} \code{vector}, the identifiers used
-
     identifier = function(value) {
       if (missing(value)) {
         colnames(private$ids)
@@ -385,23 +345,15 @@ Index <- R6::R6Class(
         private$.hide_warnings <- value
       }
     }
-
   ),
-
   private = list(
-
     elements = NULL,
     ids = NULL,
-
     confirm_default = FALSE,
     missing_default = NA,
     .hide_warnings = FALSE,
-
-    check_input = function(
-      identifier = NULL, confirm = NULL, ids = NULL, missing_identifier = NULL,
-      logical = NULL
-    ) {
-
+    check_input = function(identifier = NULL, confirm = NULL, ids = NULL, missing_identifier = NULL,
+                           logical = NULL) {
       ### check 'identifier' input
       if (!is.null(identifier)) {
         checkmate::assert_character(identifier, any.missing = FALSE)
@@ -426,11 +378,10 @@ Index <- R6::R6Class(
       if (!is.null(missing_identifier)) {
         checkmate::assert_flag(missing_identifier, na.ok = TRUE)
       }
-
     },
-
-    check_identifier_known = function(identifier) {
+    check_identifier_known = function(identifier, logical) {
       checkmate::assert_character(identifier, any.missing = FALSE, min.len = 1)
+      checkmate::assert_choice(logical, c("and", "or"))
       identifier_translated <- names(private$translate_identifier(identifier))
       unknown <- which(!identifier_translated %in% c("all", self$identifier))
       if (length(unknown) > 0) {
@@ -438,22 +389,36 @@ Index <- R6::R6Class(
           warning(
             paste0(
               "I do not know the identifier(s) '",
-              paste(identifier[unknown], collapse = "', '"),
-              "' and hence I will ignore them."
+              paste(identifier[unknown], collapse = "', '"), "'."
             ),
-            call. = FALSE, immediate. = TRUE
+            call. = FALSE, immediate. = FALSE
           )
         }
-        identifier <- identifier[-unknown]
+        if (logical == "and") {
+          if (!self$hide_warnings) {
+            warning(
+              "Because 'logical = \"and\", no element is selected.",
+              call. = FALSE, immediate. = FALSE
+            )
+          }
+          return(integer())
+        } else if (logical == "or") {
+          if (!self$hide_warnings) {
+            warning(
+              "Because 'logical = \"or\", I will ignore them.",
+              call. = FALSE, immediate. = FALSE
+            )
+          }
+          identifier <- identifier[-unknown]
+        } else {
+          unknown_error()
+        }
       }
       return(identifier)
     },
-
-    user_confirm = function(
-      action = character(), identifier = character(), ids = integer(),
-      missing_identifier = self$missing_identifier, complete = TRUE,
-      logical = "and"
-    ) {
+    user_confirm = function(action = character(), identifier = character(), ids = integer(),
+                            missing_identifier = self$missing_identifier, complete = TRUE,
+                            logical = "and") {
       checkmate::assert_flag(complete)
       cat("You are about to", action)
       if (any(identifier == "all")) {
@@ -482,7 +447,7 @@ Index <- R6::R6Class(
             } else if (logical == "or") {
               cat("with at least one of these identifiers:\n")
             } else {
-              stop("error")
+              unknown_error()
             }
           }
           identifier_bool <- private$translate_identifier(identifier)
@@ -501,7 +466,6 @@ Index <- R6::R6Class(
         return(invisible(self))
       }
     },
-
     add_identifier = function(identifier, missing_identifier) {
       checkmate::assert_character(identifier, any.missing = FALSE, min.len = 1)
       stopifnot(length(intersect(identifier, self$identifier)) == 0)
@@ -513,7 +477,6 @@ Index <- R6::R6Class(
         private$ids[, identifier] <- missing_identifier
       }
     },
-
     add_element = function(x, identifier, missing_identifier) {
       identifier_bool <- private$translate_identifier(identifier)
       new_identifier <- setdiff(names(identifier_bool), self$identifier)
@@ -525,11 +488,9 @@ Index <- R6::R6Class(
       private$ids[id, ] <- self$missing_identifier
       private$ids[id, names(identifier_bool)] <- identifier_bool
     },
-
     next_id = function() {
       length(private$elements) + 1
     },
-
     get_element = function(identifier, ids, id_names, logical) {
       checkmate::assert_flag(id_names)
       if (length(identifier) > 0 && length(ids) > 0) {
@@ -546,7 +507,6 @@ Index <- R6::R6Class(
         names = if (id_names) ids
       )
     },
-
     translate_identifier = function(identifier) {
       identifier_translated <- logical(0)
       for (value in identifier) {
@@ -558,14 +518,16 @@ Index <- R6::R6Class(
       }
       return(identifier_translated)
     },
-
     complete_identifier_bool = function(identifier_bool, missing_identifier) {
       identifier_bool[setdiff(self$identifier, names(identifier_bool))] <-
         missing_identifier
       return(identifier_bool)
     },
-
     get_ids = function(identifier, logical) {
+      ### check for unknown identifier
+      identifier <- private$check_identifier_known(
+        identifier = identifier, logical = logical
+      )
 
       ### check for identifier and inverse identifier
       for (i in identifier) {
@@ -612,9 +574,7 @@ Index <- R6::R6Class(
         identifier_bool
       )
       sort(unique(which(ids)))
-
     },
-
     merge_ids = function(..., logical) {
       ids <- list(...)
       if (logical == "and") {
@@ -622,7 +582,7 @@ Index <- R6::R6Class(
       } else if (logical == "or") {
         ids <- unique(unlist(ids))
       } else {
-        stop("error")
+        unknown_error()
       }
       checkmate::assert_integerish(ids, lower = 1, any.missing = FALSE)
       if (length(ids) == 0) {
@@ -631,41 +591,5 @@ Index <- R6::R6Class(
         sort(ids)
       }
     }
-
   )
 )
-
-#' Merge named lists
-#'
-#' @description
-#' This function merges \code{list}s based on their element names. Elements are
-#' only included in the final output \code{list}, if no former \code{list} has
-#' contributed an element with the same name.
-#'
-#' @param ...
-#' One or more named \code{list}(s).
-#'
-#' @return
-#' A \code{list}.
-#'
-#' @examples
-#' merge_lists(list("a" = 1, "b" = 2), list("b" = 3, "c" = 4))
-#'
-#' @export
-
-merge_lists <- function(...) {
-  inputs <- list(...)
-  lapply(inputs, function(input) checkmate::assert_list(input, names = "unique"))
-  final <- list()
-  for (input in inputs) {
-    for (element in names(input)) {
-      if (!element %in% names(final)) {
-        final[[element]] <- input[[element]]
-      }
-    }
-  }
-  return(final)
-}
-
-
-