+++ tests/testthat/test-storage_helpers.R
@@ -1,65 +1,46 @@
-test_that("basic Index example works", {
-  my_index <- Index$new()
-  checkmate::expect_r6(my_index, "Index")
-  my_index$
+test_that("basic Storage example works", {
+  my_Storage <- Storage$new()
+  checkmate::expect_r6(my_Storage, "Storage")
+  my_Storage$
     add(42, c("number", "rational"))$
     add(pi, c("number", "!rational"))$
     add("fear of black cats", c("text", "!rational"))$
     add("wearing a seatbelt", c("text", "rational"))$
     add(mean, "function")
   expect_equal(
-    my_index$get("rational"),
+    my_Storage$get("rational"),
     list(42, "wearing a seatbelt")
   )
   expect_length(
-    my_index$get("!rational"),
+    my_Storage$get("!rational"),
     2
   )
   expect_equal(
-    my_index$get(c("text", "!rational")),
+    my_Storage$get(c("text", "!rational")),
     list("fear of black cats")
   )
   expect_length(
-    my_index$get(ids = 4:5, id_names = TRUE),
+    my_Storage$get(ids = 4:5, id_names = TRUE),
     2
   )
   expect_equal(
-    my_index$get(c("text", "!text")),
+    my_Storage$get(c("text", "!text")),
     list()
   )
   expect_equal(
-    my_index$indices(c("text", "!text"), logical = "or"),
+    my_Storage$indices(c("text", "!text"), logical = "or"),
     1:5
   )
   expect_equal(
-    my_index$indices(c("all", "!rational"), logical = "or"),
+    my_Storage$indices(c("all", "!rational"), logical = "or"),
     1:5
   )
   expect_equal(
-    my_index$indices(identifier = c("all", "!rational"), logical = "and"),
+    my_Storage$indices(identifier = c("all", "!rational"), logical = "and"),
     2:3
   )
   expect_equal(
-    my_index$remove(ids = c(2, 4:5), shift_ids = TRUE)$get("all"),
+    my_Storage$remove(ids = c(2, 4:5), shift_ids = TRUE)$get("all"),
     list(42, "fear of black cats")
-  )
-})
-
-test_that("lists can be merged", {
-  expect_equal(
-    merge_lists(list("a" = 1, "b" = 2)),
-    list(a = 1, b = 2)
-  )
-  expect_equal(
-    merge_lists(list("a" = 1, "b" = 2), list()),
-    list(a = 1, b = 2)
-  )
-  expect_equal(
-    merge_lists(list("a" = 1, "b" = 2), list("b" = 3, "c" = 4)),
-    list(a = 1, b = 2, c = 4)
-  )
-  expect_equal(
-    merge_lists(list("a" = 1, "b" = 2), list("b" = 3, "c" = 4), list("d" = 5)),
-    list(a = 1, b = 2, c = 4, d = 5)
   )
 })